Restore GEM/AMASS demo integration and T800 multi-sample UX fixes.
Re-add GEM video import, AMASS NPZ loading, robot click-to-select for multi-sample runs, Plan A FK/IK split for parallel retarget, and constraint UI index dtype fix.

+++ kimodo/kimodo/viz/constraint_ui.py
@@ -461,8 +461,10 @@ class EEJointsKeyframeSet(ConstraintSet):
 
         constrained_idx = [self.skeleton.root_idx] + ee_joint_indices
 
-        constrained_idx = np.array(constrained_idx)
-        constrained_bone_idx = np.array(constrained_bone_idx)
+        # Force integer dtype: a single-joint end-effector leaves bone indices empty, and
+        # ``np.array([])`` defaults to float64 — numpy rejects that as a fancy index.
+        constrained_idx = np.array(constrained_idx, dtype=np.int64)
+        constrained_bone_idx = np.array(constrained_bone_idx, dtype=np.int64)
 
         # create skeleton to visualize the full-body constraint
         joint_color = np.full((self.skeleton.nbjoints, 3), (220, 220, 220))
